// src/app/bean/models/comunications.bean.ts
export class ComunicationsRS {

    code?: string; // IDX_ID
    originType?: string; // DAT_TIP_ORIGEN
    date?: Date; // DAT_FECHA
    centerCode?: string; // DAT_CENTRO_COD
    centerDesc?: string; // DAT_CENTRO_DESC
    serviceDoctorCode?: string; // DAT_SERVICIO_MEDICO_COD
    serviceDoctorDesc?: string; // DAT_SERVICIO_MEDICO_DESC
    userID?: string; // DAT_USUARIO_ID
    userDesc?: string; // DAT_USUARIO_DESC
    read?: string; // DAT_LEIDO
    readDate?: Date; // DAT_LEIDO_FECHA
    readCenterCode?: string; // DAT_LEIDO_CENTRO_COD
    readCenterDesc?: string; // DAT_LEIDO_CENTRO_DESC
    readServiceDoctorCode?: string; // DAT_LEIDO_SERVICIO_MEDICO_COD
    readServiceDoctorDesc?: string; // DAT_LEIDO_SERVICIO_MEDICO_DESC
    readUserID?: string; // DAT_LEIDO_USUARIO_ID
    readUserDesc?: string; // DAT_LEIDO_USUARIO_DESC
    requestgroupID?: string; // DAT_REQUESTGROUP_ID
    cipa?: string; // DAT_CIPA
    patientName?: string; // DAT_PACIENTE_NOMBRE
    textMessage?: string; // DAT_TEXTO

}