// src/app/bean/models/notification-message.bean.ts
export class NotificationMessageRS {

    code?: string; // IDX_ID
    prescriptorDNI?: string; // DAT_PRESCRIPTOR_DNI
    prescriptorName?: string; // DAT_PRESCRIPTOR_NOMBRE
    originServiceCode?: string; // DAT_SERVICIO_ORIGEN_COD
    originServiceDesc?: string; // DAT_SERVICIO_ORIGEN_DESC
    notificationType?: string; // DAT_TIPO_NOTIFICACION
    originEstablishment?: string; // DAT_ESTABLECIMIENTO_ORIGEN
    centerDesc?: string; // DAT_CENTRO_DESC
    requestGroup?: string; // DAT_REQUEST_GROUP
    cipa?: string; // DAT_CIPA
    patientName?: string; // DAT_PACIENTE_NOMBRE
    read?: string; // DAT_LEIDO
    readUserID?: string; // DAT_LEIDO_USUARIO_ID
    readUserDesc?: string; // DAT_LEIDO_USUARIO_DESC
    readCenterCode?: string; // DAT_LEIDO_CENTRO_COD
    readCenterDesc?: string; // DAT_LEIDO_CENTRO_DESC  
    readServiceDoctorCode?: string; // DAT_LEIDO_SERVICIO_MEDICO_COD
    readServiceDoctorDesc?: string; // DAT_LEIDO_SERVICIO_MEDICO_DESC
   

}